rename folder, default to mdx page, remove initial tag
